backend/requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
google-cloud-firestore
python-dotenv
google-generativeai
slowapi
